extract firebase request
